Use GraphicsCamera::lookAt() in BuilderWindow::centerCamera()
This changes the BuilderWindow::centerCamera() method to
use the new GraphicsCamera::lookAt() method which sets the
camera's focus to the current molecule's center instead of
moving the molecule to the origin and then looking at the
origin. This allows for the camera to be centered on the
molecule without modifying the molecule.