Replace usages of qFuzzyCompare for equality testing
This replaces the usage of qFuzzyCompare for testing  the
equality of numbers in vectors and matricies with a new
method using std::abs() and std::numeric_limits<T>::epsilon().
The qFuzzyCompare() method is only provided for float and
double arguments where as the new method works for all numeric
types.