Add Polymer and PolymerChain classes
This adds two new classes called Polymer and PolymerChain
to the chemkit library. The Polymer class represents a
generic polymer (protein, nucleic acid, etc.) and the
PolymerChain class represents a single chain in a polymer.

The Polymer class merges the functionality of the Protein
and NucleicAcid classes together and will replace them.