Add constructor with format name to MoleculeFileFormatAdaptor
This adds another constructor to the MoleculeFileFormatAdaptor<T>
class with T = LineFormat which takes a format name to use for
the adapted file format. This allows for line formats such as SMILES
to register with a different name for their files (e.g. 'smi'). This
fixes a bug with chemkit-grep when searching through a smi file.