Fix bug in DelaunayTriangulation::location()
This fixes a bug in which the DelaunayTriangulation::location()
method would sometimes get stuck in an infinate loop. Now there
is a second search performed linearly through all the tetrahedra
if we failed to find the containing tetrahedron by walking
through the delaunay structure.

Also adds a test case with a molecule from the MMFF validation
suite which was affected by this bug.