Fix build error in ChemkitConfig.cmake
This fixes a build error in which the CMakeCache.txt file
was not being created before the ChemkitConfig.cmake file
was configured and thus the ChemkitBuildTreeSettings file
was not being included correctly.

Now, a CMake variable (CHEMKIT_IN_BUILD_TREE) is defined
in the toplevlel CMakeLists.txt and is used to detect if
we are building chemkit instead of checking for the
existance of the CMakeCache.txt file.