Fix bug in the atomic-number atom typer
This fixes a bug in the atomic-number atom typer in which
the atomic number (which is stored as an unsigned char) was
being interpreted by boost::lexical_cast() as an ASCII
character rather than as a number.

To fix the issue the atomic number now converted first to an
int and then passed to boost::lexical_cast(). The test case
was also updated to verify the correct behavior.

Change-Id: I0afcc43a15cfb22c1dc81450dd4660b9005ad9e5