Fix bug in which JSON strings were not sanitized
This fixes a bug in which the ChemJsonFileFormat class would
write invalid JSON documents when the input molecule's name
contains double-quotes. Now, the strings are sanitized before
being inserted into the JSON document.

Also includes a test to verify that valid Chemical JSON files
are generated even when the input molecule's name contains
double-quotes.

Thanks to Ian Daniher for reporting this bug!

Change-Id: I90ba75b7545a5394c457ef2450834d0e280a0593