Add "--no-optimization" flag to chemkit-gen3d
This adds a new command line option to the chemkit-gen3d
program which instructs it to skip geometry optimization
when generating 3D coordinates for a molecule.

This is useful in cases when geometry optimization is
very time consuming (e.g. large molecules) or when only
the output of the coordinate prediction step is needed.

Change-Id: I79eadaceb27b57a82d6d87694ca05f030f974896