Add CoordinatePredictor::eliminateCloseContacts() method
This adds a new static method to the CoordinatePredictor class
which eliminates any close contacts between atoms by adjusting
their positions so that no two atoms are within a specified
distance (by default one Angstrom) of each other.

This method can be useful in cases where existing 3D coordinates
need to be checked and cleaned up in preparation for running a
geometry optimization.

Change-Id: I7b05741c8274f70f33e5a633b6b0b330e254115d