Change GenericFile::setFileName() to return bool
This changes the GenericFile::setFileName() method to return a
bool indicating whether or not the automatic file format detection
was successful. This allows for better error handling and better
error messages to be displayed when an invalid file format is
detected from the file name.

This also adds a test to ensure the new functionality works
correctly when given a file name with an unsupported file format.

Change-Id: I3f614434ff2eb2b4f52ed3ac14f40111c90a05eb